Split executor class in preparation for kustomize executor
